split detector_functions, ans some cleanup
