+++ .github/workflows/python_test.yml
@@ -10,8 +10,6 @@ jobs:
       max-parallel: 4
       matrix:
         python-version: [3.6, 3.7]
-        # disabled 3.7 because of 'AttributeError: 'str' object has no attribute 'decode''
-        #python-version: [3.6]
 
     steps:
     - uses: actions/checkout@v1
